Updated animation and diagnostics functions

+++ make_animation.py
@@ -7,13 +7,25 @@ from structure_function import structure_function
 from matplotlib import rc
 rc('text', usetex=True)  # LaTeX labels
 
-fname = str(sys.argv[1])
+fname = sys.argv[1]
+n_start = int(sys.argv[2]) if len(sys.argv) >= 2 else 0
+do_delete = int(sys.argv[3]) if len(sys.argv) == 3 else 0
+path = '/media/zade/Seagate Expansion Drive/Summer_Project_2019/'
+filename = path + 'figs/' + fname
 mhd = 'mhd' in fname or 'cgl' in fname
-max_n = 200
-filename = '/media/zade/Seagate Expansion Drive/Summer_Project_2019/'
-filename += 'figs/' + fname
+max_n = len(glob.glob(path+fname+'/*.athdf'))
 
-for n in range(max_n+1):
+print('Reading images from', filename)
+if do_delete:
+    print('Deleting images after')
+
+if input('Do you want to continue? (Y/N) ')[0].lower() == 'n':
+    print('Cancelling')
+    sys.exit(0)
+
+
+# Generate png images
+for n in range(n_start, max_n+1):
     print('n =', n)
     structure_function(fname, n, do_mhd=mhd)
 
@@ -46,9 +58,13 @@ def animate_mhd():
                     duration=0.2)
 
 
+# Glue together outputted image files
 animate_struct()
 if mhd:
     animate_mhd()
 
-for image in glob.glob(filename + '/*.png'):
-    os.remove(image)
+# Cleanup
+if do_delete:
+    for image in glob.glob(filename + '/*.png'):
+        os.remove(image)
+    os.rmdir(filename)
